+++ Poketochip/Projects/Data/Sources/DTO/ExampleRequestDTO.swift
@@ -1,0 +1,13 @@
+//
+//  ExampleRequestDTO.swift
+//  Infrastructure
+//
+//  Created by cha_nyeong on 5/19/24.
+//
+
+import Foundation
+
+struct ExampleRequestDTO: Encodable {
+    let name: String?
+    let id: Int?
+}
