+++ Poketochip/Projects/Data/Sources/DTO/ExampleResponseDTO.swift
@@ -1,0 +1,13 @@
+//
+//  ExampleResponseDTO.swift
+//  Infrastructure
+//
+//  Created by cha_nyeong on 5/19/24.
+//
+
+import Foundation
+
+struct ExampleResponseDTO: Decodable {
+    let name: String
+    let url: String
+}
