+++ Poketochip/Projects/Data/Sources/Repository/NetworkRepository.swift
@@ -1,0 +1,16 @@
+//
+//  NetworkRepository.swift
+//  Data
+//
+//  Created by cha_nyeong on 5/20/24.
+//
+
+import Foundation
+
+public final class NetworkRepository: baseNetworkRepository {
+    
+}
+
+public protocol baseNetworkRepository {
+    
+}
