--- Poketochip/Projects/Infrastructure/Sources/Network/RequestModel/ExampleRequest.swift
@@ -1,13 +1,0 @@
-//
-//  ExampleRequest.swift
-//  Infrastructure
-//
-//  Created by cha_nyeong on 5/16/24.
-//
-
-import Foundation
-
-struct ExampleRequestModel: Encodable {
-    let id: String
-    let name: String
-}
