--- Poketochip/Projects/Infrastructure/Sources/Network/ResponseModel/ExampleResponse.swift
@@ -1,12 +1,0 @@
-//
-//  ExampleResponse.swift
-//  Infrastructure
-//
-//  Created by cha_nyeong on 5/16/24.
-//
-
-import Foundation
-
-struct ExampleResponseModel: Decodable {
-    
-}
